api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "Auth, rate-limit, and claims-forwarding gateway for paperclipai"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "bcrypt>=4.1.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-ulid>=3.0.0",
    "redis>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "structlog>=24.0.0",
    "uvicorn[standard]>=0.30.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "testcontainers[postgres,redis]>=4.8.0",
]
